feat(authentication): magiclink handler WIP
