+++ modules/authentication/src/config/magiclink.config.ts
@@ -1,0 +1,13 @@
+export default {
+  magic_link: {
+    enabled: {
+      format: 'Boolean',
+      default: false,
+    },
+    redirect_uri: {
+      format: 'String',
+      default: '',
+      optional: true,
+    },
+  },
+};
