+++ modules/authentication/src/handlers/local.ts
@@ -166,132 +166,6 @@ export class LocalHandlers implements IAuthenticationStrategy {
       new ConduitRouteReturnDefinition('VerifyChangeEmailResponse', 'String'),
       this.verifyChangeEmail.bind(this),
     );
-
-    routingManager.route(
-      {
-        path: '/hook/passwordless-login',
-        action: ConduitRouteActions.POST,
-        description: `A webhook used to passwordless login.`,
-        bodyParams: {
-          email: ConduitString.Required,
-        },
-      },
-      new ConduitRouteReturnDefinition('PasswordLessLoginResponse', 'String'),
-      this.passwordlessLogin.bind(this),
-    );
-
-    routingManager.route(
-      {
-        path: '/hook/passwordless-login/:verificationToken',
-        action: ConduitRouteActions.GET,
-        description: `A webhook used to verify a passwordless login.`,
-        urlParams: {
-          verificationToken: ConduitString.Required,
-        },
-      },
-      new ConduitRouteReturnDefinition('VerifyPasswordlessLoginResponse', {
-        userId: ConduitString.Optional,
-        accessToken: ConduitString.Optional,
-        refreshToken: config.generateRefreshToken
-          ? ConduitString.Required
-          : ConduitString.Optional,
-        message: ConduitString.Optional,
-      }),
-      this.verifyPasswordlessLogin.bind(this),
-    );
-
-    if (config.twofa.enabled) {
-      routingManager.route(
-        {
-          path: '/local/twofa',
-          action: ConduitRouteActions.POST,
-          description: `Verifies the 2FA token.`,
-          bodyParams: {
-            email: ConduitString.Required,
-            code: ConduitString.Required,
-            token: ConduitString.Optional,
-          },
-        },
-        new ConduitRouteReturnDefinition('VerifyTwoFaResponse', {
-          userId: ConduitString.Optional,
-          accessToken: ConduitString.Optional,
-          refreshToken: config.generateRefreshToken
-            ? ConduitString.Required
-            : ConduitString.Optional,
-          message: ConduitString.Optional,
-        }),
-        this.verify2FA.bind(this),
-      );
-
-      routingManager.route(
-        {
-          path: '/local/enable-twofa',
-          action: ConduitRouteActions.UPDATE,
-          description: `Enables a phone or qr based 2FA method for a user and 
-                requires their phone number in case of phone 2FA.`,
-          middlewares: ['authMiddleware'],
-          bodyParams: {
-            method: ConduitString.Required,
-            phoneNumber: ConduitString.Optional,
-          },
-        },
-        new ConduitRouteReturnDefinition('EnableTwoFaResponse', 'String'),
-        this.enableTwoFa.bind(this),
-      );
-
-      routingManager.route(
-        {
-          path: '/local/verify-qr-code',
-          action: ConduitRouteActions.POST,
-          description: `Verifies the code the user received from scanning the QR image`,
-          middlewares: ['authMiddleware'],
-          bodyParams: {
-            code: ConduitString.Required,
-          },
-        },
-        new ConduitRouteReturnDefinition('VerifyQRCodeResponse', 'String'),
-        this.verifyQrCode.bind(this),
-      );
-
-      routingManager.route(
-        {
-          path: '/local/verifyPhoneNumber',
-          action: ConduitRouteActions.POST,
-          description: `Verifies the phone number provided for the 2FA mechanism.`,
-          middlewares: ['authMiddleware'],
-          bodyParams: {
-            code: ConduitString.Required,
-          },
-        },
-        new ConduitRouteReturnDefinition('VerifyPhoneNumberResponse', 'String'),
-        this.verifyPhoneNumber.bind(this),
-      );
-
-      routingManager.route(
-        {
-          path: '/local/disable-twofa',
-          action: ConduitRouteActions.UPDATE,
-          description: `Disables the user's 2FA mechanism.`,
-          middlewares: ['authMiddleware'],
-        },
-        new ConduitRouteReturnDefinition('DisableTwoFaResponse', 'String'),
-        this.disableTwoFa.bind(this),
-      );
-
-      routingManager.route(
-        {
-          path: '/local/change-password/verify',
-          action: ConduitRouteActions.POST,
-          description: `Used to provide the 2FA token for password change.`,
-          bodyParams: {
-            code: ConduitString.Required,
-          },
-          middlewares: ['authMiddleware'],
-        },
-        new ConduitRouteReturnDefinition('VerifyChangePasswordResponse', 'String'),
-        this.verifyChangePassword.bind(this),
-      );
-    }
   }
 
   async validate(): Promise<boolean> {
@@ -732,10 +606,6 @@ export class LocalHandlers implements IAuthenticationStrategy {
       this.emailModule = this.grpcSdk.emailProvider!;
     }
 
-    if (config.passwordless_login.enabled && this.grpcSdk.isAvailable('email')) {
-      this.emailModule = this.grpcSdk.emailProvider!;
-    }
-
     if (config.local.verification.send_email && this.grpcSdk.isAvailable('email')) {
       this.registerTemplates();
     }
@@ -753,108 +623,5 @@ export class LocalHandlers implements IAuthenticationStrategy {
       .catch(() => {
         ConduitGrpcSdk.Logger.error('Internal error while registering email templates');
       });
-  }
-
-  async passwordlessLogin(call: ParsedRouterRequest): Promise<UnparsedRouterResponse> {
-    const { email } = call.request.params;
-    const user: User | null = await User.getInstance().findOne({ email: email });
-    if (isNil(user)) throw new GrpcError(status.NOT_FOUND, 'User not found');
-
-    const token: Token = await Token.getInstance().create({
-      type: TokenType.PASSWORDLESS_LOGIN_TOKEN,
-      userId: user._id,
-      token: uuid(),
-    });
-
-    await this.sendEmailForPasswordLessLogin(user, token);
-    return 'token send';
-  }
-
-  private async sendEmailForPasswordLessLogin(user: User, token: Token) {
-    const serverConfig = await this.grpcSdk.config.get('router');
-    const url = serverConfig.hostUrl;
-
-    const result = { token, hostUrl: url };
-    const link = `${result.hostUrl}/hook/authentication/passwordless-login/${result.token.token}`;
-    await this.emailModule
-      .sendEmail('PasswordlessLogin', {
-        email: user.email,
-        sender: 'no-reply',
-        variables: {
-          link,
-        },
-      })
-      .catch(e => {
-        ConduitGrpcSdk.Logger.error(e);
-      });
-    return 'Email send';
-  }
-
-  async verifyPasswordlessLogin(
-    call: ParsedRouterRequest,
-  ): Promise<UnparsedRouterResponse> {
-    const { verificationToken } = call.request.params.verificationToken;
-    const context = call.request.context;
-    if (isNil(context))
-      throw new GrpcError(status.UNAUTHENTICATED, 'No headers provided');
-
-    const clientId = context.clientId;
-    const config = ConfigController.getInstance().config;
-    const token: Token | null = await Token.getInstance().findOne({
-      type: TokenType.PASSWORDLESS_LOGIN_TOKEN,
-      token: verificationToken,
-    });
-    if (isNil(token)) {
-      throw new GrpcError(status.NOT_FOUND, 'Passwordless login token does not exist');
-    }
-    const user: User | null = await User.getInstance().findOne({
-      _id: token.userId,
-    });
-    if (isNil(user)) throw new GrpcError(status.NOT_FOUND, 'User not found');
-
-    if (user.hasTwoFA) {
-      return this.twoFa.authenticate(user);
-    }
-    await Token.getInstance()
-      .deleteMany({ userId: token.userId, type: TokenType.PASSWORDLESS_LOGIN_TOKEN })
-      .catch(e => {
-        ConduitGrpcSdk.Logger.error(e);
-      });
-
-    const [accessToken, refreshToken] = await AuthUtils.createUserTokensAsPromise(
-      this.grpcSdk,
-      {
-        userId: user._id,
-        clientId: clientId,
-        config,
-      },
-    );
-
-    if (config.setCookies.enabled) {
-      const cookieOptions = config.setCookies.options;
-      const cookies: Cookie[] = [
-        {
-          name: 'accessToken',
-          value: accessToken.token,
-          options: cookieOptions,
-        },
-      ];
-      if (!isNil(refreshToken)) {
-        cookies.push({
-          name: 'refreshToken',
-          value: refreshToken.token,
-          options: cookieOptions,
-        });
-      }
-      return {
-        result: { message: 'Successfully authenticated' },
-        setCookies: cookies,
-      };
-    }
-    return {
-      userId: user!._id.toString(),
-      accessToken: accessToken,
-      refreshToken: !isNil(refreshToken) ? refreshToken : undefined,
-    };
   }
 }
